Moved steps modals generation functions from steps.ts to steps_modals_generation
